test: Move K8sServices test helpers to separate file
[ upstream commit 880e8fb69384f903a3f975483329996d53fe0ef8 ]

This commit only moves the helper functions. No functional changes.
